refactor(effects-manager): extract GLSL shaders to .glsl files via raw-loader (#886)
style: format glsl declaration file

fix: vertex shader uniforms, glsl.d.ts format, phoenix-ng jest glsl mock

fix: copy glsl shaders to dist after tsc build

fix: inline shader code to avoid glsl loader dependency

fix: inline shader code to avoid glsl loader dependency
