+++ packages/phoenix-event-display/package.json
@@ -18,7 +18,7 @@
     "tsc": "tsc",
     "tsc:build": "tsc -p tsconfig.build.json",
     "start": "yarn tsc:build --watch",
-    "build": "rimraf ./dist && yarn tsc:build",
+    "build": "rimraf ./dist && yarn tsc:build && node -e \"require('fs').mkdirSync('dist/managers/three-manager/shaders',{recursive:true});['hover-vertex','hover-fragment'].forEach(f=>require('fs').copyFileSync('src/managers/three-manager/shaders/'+f+'.glsl','dist/managers/three-manager/shaders/'+f+'.glsl'))\"",
     "build:esm": "yarn tsc:build --module es2018 --target es5 --outDir dist/esm",
     "build:cjs": "yarn tsc:build --module commonjs --target es5 --outDir dist/cjs",
     "build:bundle": "webpack -c configs/webpack.conf.js",
