+++ packages/phoenix-event-display/src/managers/three-manager/shaders/hover-fragment.ts
@@ -1,0 +1,11 @@
+/** Fragment shader for hover outline color. */
+export default `
+uniform float opacity;
+uniform float colorR;
+uniform float colorG;
+uniform float colorB;
+
+void main() {
+  gl_FragColor = vec4(colorR, colorG, colorB, opacity);
+}
+`;
